Broke the compute sample up into multiple files + minor cleanup
