+++ src/samples/simple_compute/CMakeLists.txt
@@ -1,5 +1,7 @@
 set(RENDER_SOURCE
-        simple_compute.cpp)
+        simple_compute.cpp
+        compute_init.cpp
+        execute.cpp)
 
 add_executable(simple_compute main.cpp ${VK_UTILS_SRC} ${RENDER_SOURCE})
